change url jonrou.dip.jp -> 124.32.233.3

+++ create_xpi.ps1
@@ -1,7 +1,11 @@
 # Powershell script to create jinrou-wakamete-analyzer.xpi
 
-Remove-Item 'jinrou-wakamete-analyzer.xpi'
-Remove-Item  'jinrou-wakamete-analyzer.zip'
+if (Test-Path 'jinrou-wakamete-analyzer.xpi' ){
+    Remove-Item 'jinrou-wakamete-analyzer.xpi'
+}
+if (Test-Path 'jinrou-wakamete-analyzer.zip' ){
+    Remove-Item  'jinrou-wakamete-analyzer.zip'
+}
 
 Compress-Archive -Path 'src\*' -DestinationPath 'jinrou-wakamete-analyzer.zip'
